pyproject.toml
[tool.poetry]
name = "engine"
version = "0.1.0"
description = ""
authors = ["Christopher Schmitt <mail@christopher.gg>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.7.3"
typer = "^0.12.3"
pluggy = "^1.5.0"
pendulum = "^3.0.0"
setuptools = "^70.0.0"
pyyaml = "^6.0.1"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
ruff = "^0.4.8"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F","RUF","C901", "W", "I", "UP", "PLE", "PLR", "PLW", "B", "A", "COM", "EM", "PYI", "Q", "RET", "SIM", "ARG", "PTH", "N","ANN", "S", "BLE","FBT", "TRY", "N", "PT"]
ignore = ["ANN101"]
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts ="--cov=. --cov-report xml"